+++ coffee_managements/Models/DoUongData.cs
@@ -19,6 +19,6 @@ namespace DoAnLapTrinhMang.Models
         public string MaLoai { get; set; }
 
         [JsonPropertyName("donGia")]
-        public decimal DonGia { get; set; }
+        public int DonGia { get; set; }
     }
 }
